Add ParsedPacket named tuple for queue; extend tests
Before, the output queue of the collector received unnamed tuples with
three fields. This broke the tests and was less understandable. The new
version uses a named tuple for clarity.

The tests were adapted to the new type in the queue and are fixed.

For backwards compatibility a check of the Python version is added and
the subprocess stdout/stderr arguments are passed depending on this
version. See #18.